% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsql-package.R
\docType{package}
\name{ggsql-package}
\alias{ggsql}
\alias{ggsql-package}
\title{ggsql: Grammar of Graphics for SQL}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Allows you to write queries that combine SQL (Structured Query Language) data retrieval with visualization specifications in a single, composable syntax. The 'ggsql' package binds directly with the 'ggsql' 'Rust' library and allows you to set up readers and writers and execute queries against it. The package also offers 'knitr' and 'shiny' integration allowing the user to use 'ggsql' in both frameworks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.ggsql.org}
  \item \url{https://github.com/posit-dev/ggsql-r}
  \item Report bugs at \url{https://github.com/posit-dev/ggsql-r/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item George Stagg \email{george.stagg@posit.co} (\href{https://orcid.org/0009-0006-3173-9846}{ORCID})
  \item Teun van den Brand \email{teun.vandenbrand@posit.co} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})
}

Other contributors:
\itemize{
  \item Posit, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
