% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{ggsql_execute_sql}
\alias{ggsql_execute_sql}
\title{Execute SQL and return a data frame}
\usage{
ggsql_execute_sql(reader, sql)
}
\arguments{
\item{reader}{A \code{GgsqlReaderR6} object created by \code{\link[=duckdb_reader]{duckdb_reader()}}.}

\item{sql}{A SQL query string.}
}
\value{
A data frame.
}
\description{
Executes a plain SQL query and returns the result as a data frame.
}
