% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{ggsql_has_visual}
\alias{ggsql_has_visual}
\title{Check if a validated query has a VISUALISE clause}
\usage{
ggsql_has_visual(x)
}
\arguments{
\item{x}{A \code{ggsql_validated} object.}
}
\value{
Logical.
}
\description{
Check if a validated query has a VISUALISE clause
}
