% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{ggsql_is_valid}
\alias{ggsql_is_valid}
\title{Check if a validated query is valid}
\usage{
ggsql_is_valid(x)
}
\arguments{
\item{x}{A \code{ggsql_validated} object.}
}
\value{
Logical.
}
\description{
Check if a validated query is valid
}
