% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{ggsql_layer_count}
\alias{ggsql_layer_count}
\title{Get the number of layers}
\usage{
ggsql_layer_count(spec)
}
\arguments{
\item{spec}{A \code{GgsqlSpecR6} object returned by \code{\link[=ggsql_execute]{ggsql_execute()}}.}
}
\value{
An integer.
}
\description{
Get the number of layers
}
