% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{ggsql_metadata}
\alias{ggsql_metadata}
\title{Get spec metadata}
\usage{
ggsql_metadata(spec)
}
\arguments{
\item{spec}{A \code{GgsqlSpecR6} object returned by \code{\link[=ggsql_execute]{ggsql_execute()}}.}
}
\value{
A list with elements \code{rows}, \code{columns}, and \code{layer_count}.
}
\description{
Get spec metadata
}
