% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{ggsql_stat_data}
\alias{ggsql_stat_data}
\title{Get stat transform data as a data frame}
\usage{
ggsql_stat_data(spec, index = 1L)
}
\arguments{
\item{spec}{A \code{GgsqlSpecR6} object returned by \code{\link[=ggsql_execute]{ggsql_execute()}}.}

\item{index}{Layer index (1-based).}
}
\value{
A data frame, or \code{NULL} if no stat transform for this layer.
}
\description{
Get stat transform data as a data frame
}
