% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{ggsql_stat_sql}
\alias{ggsql_stat_sql}
\title{Get the stat transform SQL for a specific layer}
\usage{
ggsql_stat_sql(spec, index = 1L)
}
\arguments{
\item{spec}{A \code{GgsqlSpecR6} object returned by \code{\link[=ggsql_execute]{ggsql_execute()}}.}

\item{index}{Layer index (1-based).}
}
\value{
A character string, or \code{NULL}.
}
\description{
Get the stat transform SQL for a specific layer
}
