% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{ggsql_visual}
\alias{ggsql_visual}
\title{Get the VISUALISE portion of a spec}
\usage{
ggsql_visual(spec)
}
\arguments{
\item{spec}{A \code{GgsqlSpecR6} object returned by \code{\link[=ggsql_execute]{ggsql_execute()}}.}
}
\value{
A character string.
}
\description{
Get the VISUALISE portion of a spec
}
