% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{vegalite_writer}
\alias{vegalite_writer}
\title{Create a Vega-Lite writer}
\usage{
vegalite_writer()
}
\value{
A \code{Writer} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The Vega-Lite backend is being phased out of ggsql. Use \code{\link[=svg_writer]{svg_writer()}},
\code{\link[=pdf_writer]{pdf_writer()}}, or \code{\link[=hep_writer]{hep_writer()}} instead.
}
\keyword{internal}
