% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsql-package.R
\docType{package}
\name{ggsql-package}
\alias{ggsql}
\alias{ggsql-package}
\title{ggsql: A SQL extension for declarative data visualization based on the Grammar of Graphics}
\description{
ggsql allows you to write queries that combine SQL data retrieval with visualization specifications in a single, composable syntax.
}
\author{
\strong{Maintainer}: George Stagg \email{george.stagg@posit.co} (\href{https://orcid.org/0009-0006-3173-9846}{ORCID})

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})
  \item Teun van den Brand \email{teun.vandenbrand@posit.co} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})
}

Other contributors:
\itemize{
  \item Posit, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
