% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{ggsql_render}
\alias{ggsql_render}
\title{Render a spec to Vega-Lite JSON}
\usage{
ggsql_render(writer, spec)
}
\arguments{
\item{writer}{A \code{GgsqlWriterR6} object created by \code{\link[=vegalite_writer]{vegalite_writer()}}.}

\item{spec}{A \code{GgsqlSpecR6} object returned by \code{\link[=ggsql_execute]{ggsql_execute()}}.}
}
\value{
A Vega-Lite JSON string.
}
\description{
Render a spec to Vega-Lite JSON
}
