% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{vegalite_writer}
\alias{vegalite_writer}
\title{Create a Vega-Lite writer}
\usage{
vegalite_writer()
}
\value{
A \code{Writer} object.
}
\description{
This function creates a vegalite writer which is currently the only writer
type for ggsql
}
\examples{
vegalite_writer()

}
